Major upgrade... breaking changes
Abstracted the decryption algorithm away from accepting a local
X509Certificate2 with RSA to an interface based parameter with
asynchronous methods.  Based concept on Azure KeyVault IKey concept that
a key encryption key could be located offsite, in an external API or in
some sort of HSM that renders the decryption and signing functions
impossible locally.  The IKeyEncyrptionKey only supports RSA keys at
this point, but it allows for an encrypted byte array that is an
encrypted asymmetric key to be passed and "unwrapped" asynchronously.
This will allow for offsite HSM and services such as Azure Key Vault.